Add files via upload
PDBManip is a free program for editing and manipulating PDB (Protein Data Bank) files. It has a graphical user interface and it is provided as an executable file for running on Windows® operating systems. There is a bunch of free programs on the internet for the same purpose. What discriminate between PDBManip and those programs is that PDBManip uses AngelScript language which is very similar to C/C++ language, to write scripts to do our editing tasks. I know that the text-manipulation script languages such as Awk and Perl are available, and many users, uses these to do their works. However, in developing PDBManip, I have customized scripting language, by adding predefined PDB related variables and functions, in such a way that a user with a very basic knowledge of programming can easily write very eligible scripts, even for very complicated editing and manipulating tasks. In this way, proficiency in computer programming, that is a must in using Awk or Perl, is not a barrier.